Usersettings extracted to sep file
